+++ src/PimEnterprise/Bundle/EnrichBundle/Resources/public/js/product/form/categories.js
@@ -6,15 +6,18 @@
  */
 define(
     [
+        'underscore',
         'pim/product-edit-form/categories',
         'pim/form'
     ],
-    function (Categories, BaseForm) {
+    function (_, Categories, BaseForm) {
         return Categories.extend({
             configure: function () {
                 this.trigger('tab:register', {
                     code: this.code,
-                    isVisible: _.bind(function () { return this.getFormData().meta.is_owner }, this),
+                    isVisible: function () {
+                        return this.getFormData().meta.is_owner;
+                    }.bind(this),
                     label: _.__('pim_enrich.form.product.tab.categories.title')
                 });