+++ src/PimEnterprise/Bundle/EnrichBundle/Resources/public/js/product/form/save.js
@@ -8,21 +8,19 @@ define(
     [
         'jquery',
         'underscore',
-        'pim/product-edit-form/save',
-        'pim/fetcher-registry'
+        'pim/product-edit-form/save'
     ],
-    function ($, _, Save, FetcherRegistry) {
+    function ($, _, Save) {
         return Save.extend({
             render: function () {
-                var categories = this.getFormData().categories;
                 var isOwner = this.getFormData().meta.is_owner;
 
                 if (!isOwner) {
                     this.updateSuccessMessage = _.__('pimee_enrich.entity.product_draft.info.update_successful');
                     this.updateFailureMessage = _.__('pimee_enrich.entity.product_draft.info.update_failed');
 
-                    if ('save-buttons' in this.parent.extensions) {
-                        var buttons = this.parent.extensions['save-buttons'].model.get('buttons');
+                    if (this.parent.getExtension('save-buttons')) {
+                        var buttons = this.parent.getExtension('save-buttons').model.get('buttons');
                         var saveButton = _.findWhere(buttons, {className: 'save-product'});
                         if (saveButton) {
                             saveButton.label = _.__('pimee_enrich.entity.product.btn.save_draft');