+++ src/PimEnterprise/Bundle/EnrichBundle/Resources/public/js/product/form/send-for-approval.js
@@ -11,6 +11,7 @@ define(
         'underscore',
         'backbone',
         'module',
+        'routing',
         'oro/mediator',
         'oro/messenger',
         'pim/form',
@@ -22,6 +23,7 @@ define(
         _,
         Backbone,
         module,
+        Routing,
         mediator,
         messenger,
         BaseForm,
@@ -98,7 +100,7 @@ define(
                 mediator.trigger('pim_enrich:form:state:confirm', {
                     message: this.confirmationMessage,
                     title: this.confirmationTitle,
-                    action: _.bind(this.submitDraft, this)
+                    action: this.submitDraft.bind(this)
                 });
             },
 
@@ -112,8 +114,8 @@ define(
                         {productId: this.getProductId()}
                     )
                 )
-                .then(ProductManager.generateMissing)
-                .then(_.bind(function (product) {
+                .then(ProductManager.generateMissing.bind(ProductManager))
+                .then(function (product) {
                     this.setData(product);
 
                     mediator.trigger('pim_enrich:form:entity:post_fetch', product);
@@ -122,7 +124,7 @@ define(
                         'success',
                         _.__('pimee_enrich.entity.product_draft.flash.sent_for_approval')
                     );
-                }, this))
+                }.bind(this))
                 .fail(function () {
                     messenger.notificationFlashMessage(
                         'error',